feat(headers.json): add support for env variables placeholders
